pseudo threading
Previously, you could send a command to the motors, and it will process
that command, and wait for that command to finish before reading in
another command. Now, we want to process other commands while the motor
is running. For example, you can send a STOP or PAUSE command to the
motors. Or you can POLL the encoder to take current angle reading that
would be processed by the ASCOM driver to determine error or speed, or
another POLL function that will only get called after some motor command
stops, for automatic error detection/correction.

That means we'll need to be able to run the motor commands in the
background.. I'll have to modify the serialEvent code to parse priority
commands, and only buffer motor commands