I may have fixed all the limit override bugs
At the end of the file is a comment block. This code successfully tests
the altitude limit switches and the override feature. I am assuming the
debounce latches still work, and that because there are no feedback
loops, there are no harmonics. It might just absolutely do what it's
supposed to do.